Youtube node added to movies :boom:
Take in account that all the keys that have '$' in the name were replaced by '_'. This was a workaround because a mongo limitation.

Was tested by hand. This means, could fail.
After a search on youtube if the movie doesn't have thumbnails or one of the nodes specified by @josuef in issue #5 fails, the youtube data is empty.

closes #5